First version
Based on `drmaciver/minithesis#19ecbc7b02571fbfbc7df4114dee265a77b4420e`

More effort was given to porting the Python code and making it work than
to the Elm code being idiomatic.

The `Result (Stop, TestCase) a` type is something like the IO monad in
the Haskell world - anything that wants to be able to run tests must
have it. (So, here notably the shrinkers, which in the Python version do
"run the tests with this fabricated history of random values and thread
the best solution in this state we pass wround" side-effects all over
the place.)

Shrinkers implemented:
- simplified versions of the "Delete chunk / Replace chunk with zeros"
shrinkers -- simplified so that they can be written in a "pure" way that
doesn't need to care about the results of previous runs.
- minimizing a single value through a binary search. This is a
"non-pure" shrinker: based on pass/fail status of the previous runs it
goes either up/down in the binary search process.